[MOV] website_event_track_location: merge into website_event_track
This feature was added as a dedicated module because it was added in
stable version 19.3 . We can now merge it directly into track module
where it belongs.

Task-6515688
