+++ addons/event/views/res_config_settings_views.xml
@@ -20,13 +20,6 @@
                                     </div>
                                 </div>
                                 <div class="mt-3 d-flex" invisible="not module_website_event_track">
-                                    <field name="module_website_event_track_location_display" class="w-auto"/>
-                                    <div>
-                                        <label string="Location Display" for="module_website_event_track_location_display"/><br/>
-                                        <span class="text-muted">Show the live schedule on dedicated location screens</span>
-                                    </div>
-                                </div>
-                                <div class="mt-3 d-flex" invisible="not module_website_event_track">
                                     <field name="module_website_event_track_quiz" class="w-auto"/>
                                     <div>
                                         <label string="Event Gamification" for="module_website_event_track_quiz"/><br/>
