+++ addons/website_event_track/controllers/__init__.py
@@ -1,6 +1,4 @@
-# -*- coding: utf-8 -*-
-# Part of Odoo. See LICENSE file for full copyright and licensing details.
-
 from . import event
 from . import event_track
+from . import location_display
 from . import webmanifest
