+++ addons/website_event_track/models/event_event.py
@@ -21,6 +21,13 @@ class EventEvent(models.Model):
     tracks_tag_ids = fields.Many2many(
         'event.track.tag', relation='event_track_tags_rel', string='Track Tags',
         compute='_compute_tracks_tag_ids', store=True)
+    location_display_background = fields.Image("Location Display Background")
+    location_display_upcoming_track_count = fields.Selection(
+        selection=[('2', "2"), ('3', "3"), ('4', "4"), ('5', "5")],
+        string="Coming Up Sessions",
+        default='3',
+        required=True,
+    )
 
     def _compute_track_count(self):
         data = self.env['event.track']._read_group([('stage_id.is_cancel', '!=', True)], ['event_id'], ['__count'])
@@ -58,6 +65,15 @@ class EventEvent(models.Model):
     # ------------------------------------------------------------
     # BUSINESS METHODS
     # ------------------------------------------------------------
+
+    def action_view_track_locations(self):
+        self.ensure_one()
+        action = self.env['ir.actions.actions']._for_xml_id('website_event_track.action_event_track_location')
+        action.update({
+            'name': self.env._("Track Locations"),
+            'domain': [('id', 'in', self.track_ids.location_id.ids)],
+        })
+        return action
 
     def _has_published_track(self):
         self.ensure_one()
