+++ addons/website_event_track/tests/__init__.py
@@ -2,6 +2,7 @@
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
 from . import test_event_menus
+from . import test_location_display
 from . import test_mail_features
 from . import test_track_internals
 from . import test_website_event
