+++ addons/website_event_track/views/event_track_location_views.xml
@@ -24,6 +24,7 @@
             <list string="Event Location" editable="bottom">
                 <field name="sequence" widget="handle"/>
                 <field name="name" placeholder='e.g. "Main Conference Room"'/>
+                <field name="location_display_url" widget="url"/>
             </list>
         </field>
     </record>
