--- addons/website_event_track_location_display/__init__.py
@@ -1,2 +1,0 @@
-from . import controllers
-from . import models
