--- addons/website_event_track_location_display/controllers/__init__.py
@@ -1,1 +1,0 @@
-from . import location_display
