--- addons/website_event_track_location_display/models/__init__.py
@@ -1,2 +1,0 @@
-from . import event_event
-from . import event_track_location
