--- addons/website_event_track_location_display/tests/__init__.py
@@ -1,1 +1,0 @@
-from . import test_location_display
