--- addons/website_event_track_location_display/views/event_track_location_views.xml
@@ -1,36 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-    <record id="event_event_view_form" model="ir.ui.view">
-        <field name="name">event.event.form.inherit.location.display</field>
-        <field name="model">event.event</field>
-        <field name="inherit_id" ref="event.view_event_form"/>
-        <field name="arch" type="xml">
-            <div name="button_box" position="inside">
-                <button name="action_view_track_locations" type="object" class="oe_stat_button" icon="location_on">
-                    <div class="o_stat_info">
-                        <span class="o_stat_text">Track Locations</span>
-                    </div>
-                </button>
-            </div>
-            <field name="badge_image" position="after">
-                <field name="location_display_background"/>
-                <field name="location_display_upcoming_track_count"/>
-            </field>
-        </field>
-    </record>
-
-    <record id="view_event_location_list_from_event" model="ir.ui.view">
-        <field name="name">event.track.location.list.location.display.from.event</field>
-        <field name="model">event.track.location</field>
-        <field name="inherit_id" ref="website_event_track.view_event_location_tree"/>
-        <field name="mode">primary</field>
-        <field name="arch" type="xml">
-            <xpath expr="//list" position="attributes">
-                <attribute name="create">false</attribute>
-            </xpath>
-            <xpath expr="//field[@name='name']" position="after">
-                <field name="location_display_url" widget="url"/>
-            </xpath>
-        </field>
-    </record>
-</odoo>
